pyproject.toml
[project]
name = "mlops-lstm-stock-prediction"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.20.0",
    "alembic>=1.14.0",
    "asyncpg>=0.30.0",
    "bcrypt>=4.0.0,<5.0.0",
    "fastapi>=0.124.0",
    "mlflow>=3.7.0",
    "pandera>=0.27.0",
    "passlib[bcrypt]>=1.7.4",
    "prometheus-client>=0.23.1",
    "pydantic[email]>=2.12.5",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.9",
    "sqlalchemy>=2.0.0",
    "structlog>=25.5.0",
    "torch>=2.9.1",
    "uvicorn>=0.38.0",
    "yfinance>=0.2.66",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pre-commit>=4.5.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.14.8",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow running tests",
]
addopts = [
    "--verbose",
    "--strict-markers",
    "--tb=short",
]
